requirements.txt
httpx>=0.27.0
pyyaml>=6.0
dbt-core>=1.7.0
dbt-sqlite>=1.7.0